+++ scipy/fft/tests/test_czt.py
@@ -5,7 +5,7 @@ A unit test module for czt.py
 '''
 from numpy.testing import (run_module_suite, assert_, assert_allclose,
                            assert_raises, dec)
-from scipy.fftpack.czt import (czt, zoomfft, czt_points, CZT, ZoomFFT)
+from scipy.fft.czt import (czt, zoomfft, czt_points, CZT, ZoomFFT)
 import numpy as np
 
 fft = np.fft.fft
