+++ scipy/fftpack/__init__.py
@@ -76,26 +76,6 @@ Convolutions (:mod:`scipy.fftpack.convolve`)
    init_convolution_kernel
    destroy_convolve_cache
 
-Chirp Z-transform (:mod:`scipy.fftpack.czt`)
-============================================
-
-.. module:: scipy.fftpack.czt
-
-.. autosummary::
-   :toctree: generated/
-
-   czt
-   zoomfft
-   scaledfft
-   CZT
-   ZoomFFT
-   ScaledFFT
-
-The functions are simpler to use, but less efficient, because they internally
-generate the same chirp signals with every call.  When using the same
-transform on multiple arrays of the same length, use the function-generating
-classes instead.
-
 """
 
 
@@ -115,7 +95,6 @@ from ._basic import *
 from ._pseudo_diffs import *
 from ._helper import *
 from ._realtransforms import *
-from .czt import *
 
 # Deprecated namespaces, to be removed in v2.0.0
 from . import basic, helper, pseudo_diffs, realtransforms
